Move `ReasonsForRejectionComponent` to shared namespace
Use this component to render structured reasons for rejection in the
candidate interface when given.
