+++ app/components/candidate_interface/course_choices_review_component.rb
@@ -30,6 +30,7 @@ module CandidateInterface
         type_row(application_choice.course),
         course_length_row(application_choice.course),
         start_date_row(application_choice),
+        rejection_reasons_row(application_choice),
       ].compact
 
       rows.tap do |r|
@@ -170,6 +171,26 @@ module CandidateInterface
         key: 'Reason for offer withdrawal',
         value: application_choice.offer_withdrawal_reason,
       }
+    end
+
+    def rejection_reasons_row(application_choice)
+      return nil unless application_choice.rejected?
+
+      if FeatureFlag.active?(:structured_reasons_for_rejection) && application_choice.structured_rejection_reasons.present?
+        {
+          key: 'Feedback',
+          value: render(ReasonsForRejectionComponent.new(
+            application_choice: application_choice,
+            reasons_for_rejection: ProviderInterface::ReasonsForRejection.new(application_choice.structured_rejection_reasons),
+            editable: false,
+          )),
+        }
+      elsif application_choice.rejection_reason.present?
+        {
+          key: 'Feedback',
+          value: application_choice.rejection_reason,
+        }
+      end
     end
 
     def has_multiple_sites?(application_choice)
