+++ app/components/candidate_interface/rejection_reasons_component.rb
@@ -34,10 +34,17 @@ module CandidateInterface
     end
 
     def rejection_reasons_row(application_choice)
-      {
-        key: 'Feedback',
-        value: application_choice.rejection_reason,
-      }
+      if FeatureFlag.active?(:structured_reasons_for_rejection) && application_choice.structured_rejection_reasons.present?
+        {
+          key: 'Feedback',
+          value: render(ReasonsForRejectionComponent.new(application_choice: application_choice)),
+        }
+      else
+        {
+          key: 'Feedback',
+          value: application_choice.rejection_reason,
+        }
+      end
     end
   end
 end
