+++ app/views/provider_interface/reasons_for_rejection/check.html.erb
@@ -12,7 +12,7 @@
         Your feedback will be given to the candidate in the format below.
       </p>
 
-      <%= render ProviderInterface::ReasonsForRejectionComponent.new(
+      <%= render ReasonsForRejectionComponent.new(
         application_choice: @application_choice,
         reasons_for_rejection: @wizard.to_model,
         editable: true,
